+++ src/api/raw_models.ts
@@ -1,15 +1,15 @@
 import type { ID, ProjectID, SectionID, LabelID } from "./models";
 
 export const UnknownProject: IProjectRaw = {
-  id: -1,
+  id: "-1",
   parent_id: null,
   order: -1,
   name: "Unknown project",
 };
 
 export const UnknownSection: ISectionRaw = {
-  id: -1,
-  project_id: -1,
+  id: "-1",
+  project_id: "-1",
   order: -1,
   name: "Unknown section",
 };
@@ -17,22 +17,28 @@ export const UnknownSection: ISectionRaw = {
 export interface ITaskRaw {
   id: ID;
   project_id: ProjectID;
-  section_id: SectionID;
-  label_ids: LabelID[];
+  section_id: SectionID | null;
+  /**
+   * The task has direct reference to the label names, and does not reference the label objects.
+   *
+   * That is, if you have a label {id:"123abc", name:"errand"} in the labels dataset, a task with
+   * that label will have the attribute { label: ["errand"] }.
+   */
+  labels: string[];
   priority: number;
   content: string;
   order: number;
-  parent?: ID;
+  parent_id?: ID | null;
   due?: {
     recurring: boolean;
-    date: string;
-    datetime?: string;
+    date: string | null;
+    datetime?: string | null;
   };
 }
 
 export interface IProjectRaw {
   id: ProjectID;
-  parent_id?: ProjectID;
+  parent_id?: ProjectID | null;
   order: number;
   name: string;
 }