+++ tests/tasks.ts
@@ -7,23 +7,23 @@ describe("Task tree parsing", () => {
   it("Tree matches parents and subtasks", () => {
     const apiTasks: ITaskRaw[] = [
       {
-        id: 1,
+        id: "1",
         priority: 1,
         order: 0,
         content: "Parent task",
-        project_id: 0,
-        section_id: 0,
-        label_ids: [],
+        project_id: "123",
+        section_id: null,
+        labels: [],
       },
       {
-        id: 2,
+        id: "2",
         priority: 2,
         order: 1,
         content: "Subtask",
-        parent: 1,
-        project_id: 0,
-        section_id: 0,
-        label_ids: [],
+        parent_id: "1",
+        project_id: "123",
+        section_id: null,
+        labels: [],
       },
     ];
 
@@ -34,21 +34,21 @@ describe("Task tree parsing", () => {
     assert.lengthOf(parent.children, 1, "Parent task has one subtask");
 
     const child = parent.children[0];
-    assert.equal(child.id, 2, "Parent has expected subtask");
+    assert.equal(child.id, "2", "Parent has expected subtask");
     assert.equal(child.parent, parent, "Subtask has expected parent.");
   });
 
   it("Subtask without parent in scope is a top-level node", () => {
     const apiTasks: ITaskRaw[] = [
       {
-        id: 2,
+        id: "2",
         priority: 2,
         order: 1,
         content: "Subtask",
-        parent: 1,
-        project_id: 0,
-        section_id: 0,
-        label_ids: [],
+        parent_id: "1",
+        project_id: "123",
+        section_id: null,
+        labels: [],
       },
     ];
 
@@ -60,7 +60,7 @@ describe("Task tree parsing", () => {
 describe("Task date parsing", () => {
   it("Tasks with dates only have no time", () => {
     const apiTask: ITaskRaw = {
-      id: 1,
+      id: "1",
       priority: 1,
       order: 1,
       content: "Task",
@@ -69,9 +69,9 @@ describe("Task date parsing", () => {
         datetime: null,
         recurring: false,
       },
-      project_id: 0,
-      section_id: 0,
-      label_ids: [],
+      project_id: "123",
+      section_id: null,
+      labels: [],
     };
 
     const task = new Task(apiTask);
@@ -81,7 +81,7 @@ describe("Task date parsing", () => {
 
   it("Tasks with time have time", () => {
     const apiTask: ITaskRaw = {
-      id: 1,
+      id: "1",
       priority: 1,
       order: 1,
       content: "Task",
@@ -90,9 +90,9 @@ describe("Task date parsing", () => {
         datetime: "2019-09-12T02:30:00Z",
         recurring: false,
       },
-      project_id: 0,
-      section_id: 0,
-      label_ids: [],
+      project_id: "123",
+      section_id: null,
+      labels: [],
     };
 
     const task = new Task(apiTask);
@@ -104,23 +104,23 @@ describe("Task date parsing", () => {
 describe("Task comparisons", () => {
   it("Tasks use Todoist order if no ordering provided", () => {
     const first = new Task({
-      id: 1,
+      id: "1",
       priority: 1,
       order: 0,
       content: "Parent task",
-      project_id: 0,
-      section_id: 0,
-      label_ids: [],
+      project_id: "123",
+      section_id: null,
+      labels: [],
     });
 
     const second = new Task({
-      id: 2,
+      id: "2",
       priority: 2,
       order: 1,
       content: "Subtask",
-      project_id: 0,
-      section_id: 0,
-      label_ids: [],
+      project_id: "123",
+      section_id: null,
+      labels: [],
     });
 
     assert.isBelow(first.compareTo(second, []), 0);
@@ -129,23 +129,23 @@ describe("Task comparisons", () => {
 
   it("Tasks are sorted by priority (high to low)", () => {
     const lowPriority = new Task({
-      id: 1,
+      id: "1",
       priority: 1,
       order: 0,
       content: "Parent task",
-      project_id: 0,
-      section_id: 0,
-      label_ids: [],
+      project_id: "123",
+      section_id: null,
+      labels: [],
     });
 
     const highPriority = new Task({
-      id: 2,
+      id: "2",
       priority: 2,
       order: 1,
       content: "Subtask",
-      project_id: 0,
-      section_id: 0,
-      label_ids: [],
+      project_id: "123",
+      section_id: null,
+      labels: [],
     });
 
     assert.isBelow(highPriority.compareTo(lowPriority, ["priority"]), 0);
@@ -154,23 +154,23 @@ describe("Task comparisons", () => {
 
   it("Tasks fallback to Todoist ordering if priority is the same", () => {
     const first = new Task({
-      id: 1,
+      id: "1",
       priority: 2,
       order: 0,
       content: "Parent task",
-      project_id: 0,
-      section_id: 0,
-      label_ids: [],
+      project_id: "123",
+      section_id: null,
+      labels: [],
     });
 
     const second = new Task({
-      id: 2,
+      id: "2",
       priority: 2,
       order: 1,
       content: "Subtask",
-      project_id: 0,
-      section_id: 0,
-      label_ids: [],
+      project_id: "123",
+      section_id: null,
+      labels: [],
     });
 
     assert.isBelow(first.compareTo(second, []), 0);
@@ -179,7 +179,7 @@ describe("Task comparisons", () => {
 
   it("Tasks are sorted by date (earliest -> latest).", () => {
     const first = new Task({
-      id: 1,
+      id: "1",
       priority: 2,
       order: 0,
       content: "Parent task",
@@ -188,13 +188,13 @@ describe("Task comparisons", () => {
         date: "2019-09-01",
         datetime: null,
       },
-      project_id: 0,
-      section_id: 0,
-      label_ids: [],
+      project_id: "123",
+      section_id: null,
+      labels: [],
     });
 
     const second = new Task({
-      id: 2,
+      id: "2",
       priority: 2,
       order: 1,
       content: "Subtask",
@@ -203,9 +203,9 @@ describe("Task comparisons", () => {
         date: "2019-09-07",
         datetime: null,
       },
-      project_id: 0,
-      section_id: 0,
-      label_ids: [],
+      project_id: "123",
+      section_id: null,
+      labels: [],
     });
 
     assert.isBelow(first.compareTo(second, ["date"]), 0);
@@ -214,7 +214,7 @@ describe("Task comparisons", () => {
 
   it("Tasks on the same day are sorted in time order (earliest -> latest -> no time)", () => {
     const first = new Task({
-      id: 1,
+      id: "1",
       priority: 2,
       order: 0,
       content: "Parent task",
@@ -223,13 +223,13 @@ describe("Task comparisons", () => {
         date: null,
         datetime: "2019-08-01T02:30:00Z",
       },
-      project_id: 0,
-      section_id: 0,
-      label_ids: [],
+      project_id: "123",
+      section_id: null,
+      labels: [],
     });
 
     const second = new Task({
-      id: 2,
+      id: "2",
       priority: 2,
       order: 1,
       content: "Subtask",
@@ -238,13 +238,13 @@ describe("Task comparisons", () => {
         date: null,
         datetime: "2019-08-01T03:30:00Z",
       },
-      project_id: 0,
-      section_id: 0,
-      label_ids: [],
+      project_id: "123",
+      section_id: null,
+      labels: [],
     });
 
     const third = new Task({
-      id: 2,
+      id: "2",
       priority: 2,
       order: 1,
       content: "Subtask",
@@ -253,9 +253,9 @@ describe("Task comparisons", () => {
         date: "2019-08-01",
         datetime: null,
       },
-      project_id: 0,
-      section_id: 0,
-      label_ids: [],
+      project_id: "123",
+      section_id: null,
+      labels: [],
     });
 
     assert.isBelow(first.compareTo(second, ["date"]), 0);
@@ -269,7 +269,7 @@ describe("Task comparisons", () => {
 
   it("Task ordered by priority falls to the next if equal", () => {
     const first = new Task({
-      id: 1,
+      id: "1",
       priority: 1,
       order: 1,
       content: "Task",
@@ -278,13 +278,13 @@ describe("Task comparisons", () => {
         date: "2019-08-02",
         datetime: null,
       },
-      project_id: 0,
-      section_id: 0,
-      label_ids: [],
+      project_id: "123",
+      section_id: null,
+      labels: [],
     });
 
     const second = new Task({
-      id: 1,
+      id: "1",
       priority: 1,
       order: 1,
       content: "Task",
@@ -293,9 +293,9 @@ describe("Task comparisons", () => {
         date: "2019-08-03",
         datetime: null,
       },
-      project_id: 0,
-      section_id: 0,
-      label_ids: [],
+      project_id: "123",
+      section_id: null,
+      labels: [],
     });
 
     assert.isBelow(first.compareTo(second, ["priority", "date"]), 0);
@@ -304,7 +304,7 @@ describe("Task comparisons", () => {
 
   it("Task ordered by date falls to the next if equal", () => {
     const first = new Task({
-      id: 1,
+      id: "1",
       priority: 2,
       order: 1,
       content: "Task",
@@ -313,13 +313,13 @@ describe("Task comparisons", () => {
         date: "2019-08-03",
         datetime: null,
       },
-      project_id: 0,
-      section_id: 0,
-      label_ids: [],
+      project_id: "123",
+      section_id: null,
+      labels: [],
     });
 
     const second = new Task({
-      id: 1,
+      id: "1",
       priority: 1,
       order: 1,
       content: "Task",
@@ -328,9 +328,9 @@ describe("Task comparisons", () => {
         date: "2019-08-03",
         datetime: null,
       },
-      project_id: 0,
-      section_id: 0,
-      label_ids: [],
+      project_id: "123",
+      section_id: null,
+      labels: [],
     });
 
     assert.isBelow(first.compareTo(second, ["date", "priority"]), 0);